+++ src/rules/no-eol-whitespace/index.js
@@ -1,4 +1,6 @@
 import {
+  isOnlyWhitespace,
+  optionsHaveIgnored,
   report,
   ruleMessages,
   validateOptions,
@@ -11,11 +13,19 @@ export const messages = ruleMessages(ruleName, {
   rejected: "Unexpected whitespace at end of line",
 })
 
-const whitespacesToReject = [ " ", "\t" ]
+const whitespacesToReject = new Set([ " ", "\t" ])
 
-export default function (actual) {
+export default function (on, options) {
   return (root, result) => {
-    const validOptions = validateOptions(result, ruleName, { actual })
+    const validOptions = validateOptions(result, ruleName, {
+      actual: on,
+    }, {
+      optional: true,
+      actual: options,
+      possible: {
+        ignore: ["empty-lines"],
+      },
+    })
     if (!validOptions) { return }
 
     const rootString = root.toString()
@@ -24,15 +34,24 @@ export default function (actual) {
       target: [ "\n", "\r" ],
       comments: "check",
     }, match => {
-      if (whitespacesToReject.indexOf(rootString[match.startIndex - 1]) !== -1) {
-        report({
-          message: messages.rejected,
-          node: root,
-          index: match.startIndex - 1,
-          result,
-          ruleName,
-        })
+      // If the character before newline is not whitespace, ignore
+      if (!whitespacesToReject.has(rootString[match.startIndex - 1])) { return }
+
+      if (optionsHaveIgnored(options, "empty-lines")) {
+        // If there is only whitespace between the previous newline and
+        // this newline, ignore
+        const lineBefore = rootString.substring(match.startIndex + 1,
+          rootString.lastIndexOf("\n", match.startIndex - 1))
+        if (isOnlyWhitespace(lineBefore)) { return }
       }
+
+      report({
+        message: messages.rejected,
+        node: root,
+        index: match.startIndex - 1,
+        result,
+        ruleName,
+      })
     })
   }
 }